+++ packages/iocuak/src/binding/decorators/injectable.ts
@@ -4,28 +4,23 @@ import { MetadataKey } from '../../reflectMetadata/models/domain/MetadataKey';
 import { bindingScopeApiToBindingScopeMap } from '../models/api/bindingScopeApiToBindingScopeMap';
 import { InjectableOptionsApi } from '../models/api/InjectableOptionsApi';
 import { BindingScope } from '../models/domain/BindingScope';
+import { BindingTag } from '../models/domain/BindingTag';
 import { BindingType } from '../models/domain/BindingType';
 import { TypeBinding } from '../models/domain/TypeBinding';
 import { getDefaultBindingScope } from '../utils/domain/getDefaultBindingScope';
 
-export function injectable(bindingApi?: InjectableOptionsApi): ClassDecorator {
+export function injectable(options?: InjectableOptionsApi): ClassDecorator {
   // eslint-disable-next-line @typescript-eslint/ban-types
   const decorator: ClassDecorator = <TFunction extends Function>(
     target: TFunction,
   ): TFunction | void => {
-    const bindingId: ServiceId =
-      bindingApi?.id ?? (target as unknown as Newable);
-
-    const bindingScope: BindingScope =
-      bindingApi?.scope === undefined
-        ? getDefaultBindingScope()
-        : bindingScopeApiToBindingScopeMap[bindingApi.scope];
+    const bindingId: ServiceId = options?.id ?? (target as unknown as Newable);
 
     const binding: TypeBinding = {
       bindingType: BindingType.type,
       id: bindingId,
-      scope: bindingScope,
-      tags: [],
+      scope: getBindingScope(options),
+      tags: getTags(options),
       type: target as unknown as Newable,
     };
 
@@ -33,4 +28,27 @@ export function injectable(bindingApi?: InjectableOptionsApi): ClassDecorator {
   };
 
   return decorator;
+}
+
+function getBindingScope(options?: InjectableOptionsApi): BindingScope {
+  const bindingScope: BindingScope =
+    options?.scope === undefined
+      ? getDefaultBindingScope()
+      : bindingScopeApiToBindingScopeMap[options.scope];
+
+  return bindingScope;
+}
+
+function getTags(options?: InjectableOptionsApi): BindingTag[] {
+  const tagOrTags: BindingTag | BindingTag[] = options?.tags ?? [];
+
+  let tags: BindingTag[];
+
+  if (Array.isArray(tagOrTags)) {
+    tags = tagOrTags;
+  } else {
+    tags = [tagOrTags];
+  }
+
+  return tags;
 }