+++ packages/iocuak/CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 
-## [UNRELEASED]
+## 0.3.0 - 2022-06-28
 
 ### Added
 - Added `BindingTag`.